// src/types/logos.types.tsx
export interface ILogo {
  
  id: number;
  name: string;
  image: string;
  muscleGroups: string[];
  equipment: string;
  difficulty: string;
  duration: number;
  caloriesBurned: number;
  sets: number;
  reps: string;
  rating: number;
  description: string;
  instructions: string[];
};
export interface IPlanLogo extends ILogo {
  completed: boolean;
}